requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0

# CLIP encoders
open-clip-torch>=2.24.0
transformers>=4.36.0

# Vector search
faiss-cpu>=1.7.4

# Orchestration
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6

# Video processing
decord>=0.6.0
opencv-python>=4.8.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Web UI
chainlit>=1.0.0

# Data validation
pydantic>=2.0.0

# HTTP clients
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.66.0

# Evaluation
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Logging
structlog>=23.2.0

# API clients
openai>=1.6.0
anthropic>=0.8.0
google-generativeai>=0.3.0

# Optional: for tiny local LLM
# llama-cpp-python>=0.2.0

